+++ patches/command_pyroscope.cc
@@ -545,6 +545,33 @@ torrent::Object cmd_string_contains_i(rpc::target_type target, const torrent::Ob
 }
 
 
+torrent::Object cmd_value(rpc::target_type target, const torrent::Object::list_type& args) {
+    if (args.size() < 1) {
+        throw torrent::input_error("'value' takes at least a number argument!");
+    }
+    if (args.size() > 2) {
+        throw torrent::input_error("'value' takes at most two arguments!");
+    }
+
+    torrent::Object::value_type val = 0;
+    if (args.front().is_value()) {
+        val = args.front().as_value();
+    } else {
+        int base = args.size() > 1 ? args.back().is_value() ?
+                   args.back().as_value() : strtol(args.back().as_string().c_str(), NULL, 10) : 10;
+        char* endptr = 0;
+
+        val = strtoll(args.front().as_string().c_str(), &endptr, base);
+        while (*endptr == ' ' || *endptr == '\n') ++endptr;
+        if (*endptr) {
+            throw torrent::input_error("Junk at end of number: " + args.front().as_string());
+        }
+    }
+
+    return val;
+}
+
+
 // Backports from 0.9.2
 #if (API_VERSION < 3)
 template <typename InputIterator, typename OutputIterator> OutputIterator
@@ -622,6 +649,7 @@ void initialize_command_pyroscope() {
 
     CMD2_ANY_LIST("string.contains", &cmd_string_contains);
     CMD2_ANY_LIST("string.contains_i", &cmd_string_contains_i);
+    CMD2_ANY_LIST("value", &cmd_value);
     CMD2_ANY_LIST("compare", &apply_compare);
     CMD2_ANY("ui.bind_key", &apply_ui_bind_key);
     CMD2_VAR_VALUE("ui.bind_key.verbose", 1);