+++ CHANGES.md
@@ -13,6 +13,7 @@
 * Command: Added `ui.bind_key.verbose` flag (issue #55)
 * Command: Added `d.multicall.filtered`
 * Command: Added `string.contains` and `string.contains_i`
+* Command: Added `value` conversion command
 
 
 ## 2017-03-12 PS-1.0 “First version-tagged release”